refactor: do not mark sk::ffi as pub
